proxy-service - implement PrimesClient and AsyncIterator
PrimesClient is a client for our gRPC command.

nextPrimeAfter(from) is a simple method which calls the server and
returns the result as an integer.

allPrimesTo(limit) prepares an iterator to make successive calls to
nextPrimeAfter until the limit is reached.

AsyncIterator is a custom iterator which is used by allPrimesTo. It
implements the StreamObserver interface so it may also be used if
the implementation of nextPrimeAfter is moved to the server.